Move permission tests to its own set
